+++ crates/runen_spatial_streaming/src/priority.rs
@@ -3,11 +3,11 @@ use runen_spatial::ChunkCoord3;
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Default)]
 pub struct ChunkPriority {
     pub rank: u32,
-    pub distance_squared: i64,
+    pub distance_squared: i128,
 }
 
 impl ChunkPriority {
-    pub fn new(rank: u32, distance_squared: i64) -> Self {
+    pub fn new(rank: u32, distance_squared: i128) -> Self {
         Self {
             rank,
             distance_squared,
@@ -15,9 +15,9 @@ impl ChunkPriority {
     }
 }
 
-pub(crate) fn distance_squared(a: ChunkCoord3, b: ChunkCoord3) -> i64 {
-    let dx = i64::from(a.x) - i64::from(b.x);
-    let dy = i64::from(a.y) - i64::from(b.y);
-    let dz = i64::from(a.z) - i64::from(b.z);
+pub(crate) fn distance_squared(a: ChunkCoord3, b: ChunkCoord3) -> i128 {
+    let dx = i128::from(a.x) - i128::from(b.x);
+    let dy = i128::from(a.y) - i128::from(b.y);
+    let dz = i128::from(a.z) - i128::from(b.z);
     dx * dx + dy * dy + dz * dz
 }
